refactor: package py-diceware with src structure
