+++ setup.cfg
@@ -1,0 +1,26 @@
+[metadata]
+name = py-diceware
+author = TomJGooding
+description =  Diceware passphrase generator built with Python
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/TomJGooding/py-diceware
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+include_package_data = True
+install_requires =
+    importlib-resources
+
+[options.packages.find]
+where = src
+
+[options.entry_points]
+console_scripts =
+    py-diceware = py_diceware.cli:main
