+++ src/py_diceware/__main__.py
@@ -1,0 +1,4 @@
+from py_diceware.cli import main
+
+if __name__ == "__main__":
+    main()
