+++ src/py_diceware/wordlist.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+from importlib_resources import files
+
+import py_diceware.wordlists
+
+
+@dataclass
+class WordList:
+    filename: str
+    path = files(py_diceware.wordlists)
